# contracts/api/openapi.json
{
  "components": {
    "schemas": {
      "AcquisitionStart": {
        "description": "Public handoff from source acquisition to the ingestion workflow.",
        "properties": {
          "candidate": {
            "$ref": "#/components/schemas/SourceCandidate"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "ingestion_job_id": {
            "title": "Ingestion Job Id",
            "type": "string"
          },
          "ingestion_status": {
            "title": "Ingestion Status",
            "type": "string"
          }
        },
        "required": [
          "candidate",
          "ingestion_job_id",
          "ingestion_status",
          "created_at"
        ],
        "title": "AcquisitionStart",
        "type": "object"
      },
...
      "CandidateInsight": {
        "description": "Ranked candidate projection safe for an API response.",
        "properties": {
          "contributions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/FeatureContributionInsight"
            },
            "title": "Contributions",
            "type": "array"
          },
          "eligible": {
            "title": "Eligible",
            "type": "boolean"
          },
          "entity_id": {
            "title": "Entity Id",
            "type": "string"
          },
          "entity_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Name"
          },
          "failed_constraints": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Failed Constraints",
            "type": "array"
          },
          "missing_required_features": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Missing Required Features",
            "type": "array"
          },
          "rank": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank"
          },
          "score": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          }
        },
        "required": [
          "entity_id",
          "eligible"
        ],
        "title": "CandidateInsight",
        "type": "object"
      },
...
      "CopilotQueryRequest": {
        "description": "Natural-language question plus optional candidate and quality scope.",
        "properties": {
          "entityIds": {
            "default": [],
            "items": {
              "type": "string"
            },
            "maxItems": 500,
            "title": "Entityids",
            "type": "array"
          },
          "minQualityScore": {
            "default": 0.0,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Minqualityscore",
            "type": "number"
          },
          "question": {
            "maxLength": 2000,
            "minLength": 3,
            "title": "Question",
            "type": "string"
          },
          "responseLanguage": {
            "anyOf": [
              {
                "enum": [
                  "en",
                  "zh-TW"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responselanguage"
          },
          "sessionId": {
            "anyOf": [
              {
                "pattern": "^ses_[a-f0-9]{32}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sessionid"
          },
          "topicHint": {
            "anyOf": [
              {
                "pattern": "^[a-z][a-z0-9_]{0,99}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topichint"
          }
        },
        "required": [
          "question"
        ],
        "title": "CopilotQueryRequest",
        "type": "object"
      },
      "CopilotResponse": {
        "description": "Grounded answer with plan, trace, ranking, evidence, and limitations.",
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "assumptions": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Assumptions",
            "type": "array"
          },
          "candidates": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/CandidateInsight"
            },
            "title": "Candidates",
            "type": "array"
          },
          "citations": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/EvidenceCitation"
            },
            "title": "Citations",
            "type": "array"
          },
          "comparison": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EntityComparison"
              },
              {
                "type": "null"
              }
            ]
          },
          "data_requirement": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DataRequirement"
              },
              {
                "type": "null"
              }
            ]
          },
          "dataset_inspection": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DatasetInspection"
              },
              {
                "type": "null"
              }
            ]
          },
          "decomposition": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DecomposedQuery"
              },
              {
                "type": "null"
              }
            ]
          },
          "forecast_result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ForecastResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "impact_analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ImpactAnalysis"
              },
              {
                "type": "null"
              }
            ]
          },
          "limitations": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DataLimitations"
              },
              {
                "type": "null"
              }
            ]
          },
          "multi_dataset_analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MultiDatasetAnalysis"
              },
              {
                "type": "null"
              }
            ]
          },
          "observation_series": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ObservationSeries"
              },
              {
                "type": "null"
              }
            ]
          },
          "plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DecisionExecutionPlan"
              },
              {
                "type": "null"
              }
            ]
          },
          "routed_plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RoutedToolPlan"
              },
              {
                "type": "null"
              }
            ]
          },
          "session_id": {
            "anyOf": [
              {
                "pattern": "^ses_[a-f0-9]{32}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "source_candidates": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SourceCandidate"
            },
            "title": "Source Candidates",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/CopilotStatus"
          },
          "tool_trace": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ToolTrace"
            },
            "title": "Tool Trace",
            "type": "array"
          },
          "visualizations": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/VisualizationSpec"
            },
            "title": "Visualizations",
            "type": "array"
          },
          "warnings": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          },
          "web_citations": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/WebCitation"
            },
            "title": "Web Citations",
            "type": "array"
          }
        },
        "required": [
          "status",
          "answer",
          "generated_at"
        ],
        "title": "CopilotResponse",
        "type": "object"
      },
      "CopilotStatus": {
        "description": "Whether the available evidence supports an answer.",
        "enum": [
          "answered",
          "insufficient_data",
          "acquisition_required",
          "unsupported_question"
        ],
        "title": "CopilotStatus",
        "type": "string"
      },
...
      "DataRequirement": {
        "description": "A bounded description of data missing from the published catalog.",
        "properties": {
          "accepted_formats": {
            "default": [
              "csv",
              "json",
              "excel"
            ],
            "items": {
              "$ref": "#/components/schemas/FileFormat"
            },
            "title": "Accepted Formats",
            "type": "array"
          },
          "entity_ids": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Entity Ids",
            "type": "array"
          },
          "metric_codes": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Metric Codes",
            "type": "array"
          },
          "time_expression": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Expression"
          },
          "topic_terms": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Topic Terms",
            "type": "array"
          }
        },
        "title": "DataRequirement",
        "type": "object"
      },
...
      "DatasetInspection": {
        "description": "Schema and coverage selected from the published catalog.",
        "properties": {
          "additive": {
            "default": true,
            "title": "Additive",
            "type": "boolean"
          },
          "available_metrics": {
            "items": {
              "type": "string"
            },
            "title": "Available Metrics",
            "type": "array"
          },
          "dataset_id": {
            "title": "Dataset Id",
            "type": "string"
          },
          "dataset_version": {
            "title": "Dataset Version",
            "type": "string"
          },
          "entity_count": {
            "minimum": 0.0,
            "title": "Entity Count",
            "type": "integer"
          },
          "grain": {
            "items": {
              "type": "string"
            },
            "title": "Grain",
            "type": "array"
          },
          "metric_code": {
            "title": "Metric Code",
            "type": "string"
          },
          "period_end": {
            "title": "Period End",
            "type": "string"
          },
          "period_start": {
            "title": "Period Start",
            "type": "string"
          },
          "quality_score": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Quality Score",
            "type": "number"
          },
          "topic": {
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "dataset_id",
          "dataset_version",
          "topic",
          "grain",
          "metric_code",
          "available_metrics",
          "period_start",
          "period_end",
          "entity_count",
          "quality_score"
        ],
        "title": "DatasetInspection",
        "type": "object"
      },
...
      "EntityComparison": {
        "description": "Comparable changes calculated from one compatible observation series.",
        "properties": {
          "changes": {
            "items": {
              "$ref": "#/components/schemas/EntityChange"
            },
            "title": "Changes",
            "type": "array"
          },
          "metric_code": {
            "title": "Metric Code",
            "type": "string"
          },
          "unit_code": {
            "title": "Unit Code",
            "type": "string"
          }
        },
        "required": [
          "metric_code",
          "unit_code",
          "changes"
        ],
        "title": "EntityComparison",
        "type": "object"
      },
      "EvidenceCitation": {
        "description": "Public provenance without leaking local storage paths or credentials.",
        "properties": {
          "citation_id": {
            "title": "Citation Id",
            "type": "string"
          },
          "dataset_id": {
            "title": "Dataset Id",
            "type": "string"
          },
          "dataset_version": {
            "title": "Dataset Version",
            "type": "string"
          },
          "excerpt": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/EvidenceExcerptRow"
            },
            "maxItems": 100,
            "title": "Excerpt",
            "type": "array"
          },
          "quality_score": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Quality Score",
            "type": "number"
          },
          "retrieved_at": {
            "format": "date-time",
            "title": "Retrieved At",
            "type": "string"
          }
        },
        "required": [
          "citation_id",
          "dataset_id",
          "dataset_version",
          "quality_score",
          "retrieved_at"
        ],
        "title": "EvidenceCitation",
        "type": "object"
      },
...
      "FileFormat": {
        "enum": [
          "csv",
          "excel",
          "json",
          "pdf"
        ],
        "title": "FileFormat",
        "type": "string"
      },
...
      "ForecastResult": {
        "description": "Forecast values and the lineage needed to judge them.",
        "properties": {
          "evaluation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ForecastEvaluation"
              },
              {
                "type": "null"
              }
            ]
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "metric_code": {
            "title": "Metric Code",
            "type": "string"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/ForecastPoint"
            },
            "minItems": 1,
            "title": "Points",
            "type": "array"
          }
        },
        "required": [
          "metric_code",
          "model_version",
          "points",
          "generated_at"
        ],
        "title": "ForecastResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LinkAcquisitionRequest": {
        "description": "A reviewer pointing at data the manifest does not list yet.",
        "properties": {
          "submittedBy": {
            "maxLength": 320,
            "minLength": 1,
            "title": "Submittedby",
            "type": "string"
          },
          "topicHint": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topichint"
          },
          "url": {
            "maxLength": 2000,
            "minLength": 12,
            "pattern": "^https://",
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "submittedBy"
        ],
        "title": "LinkAcquisitionRequest",
        "type": "object"
      },
      "LinkAcquisitionStart": {
        "description": "Public handoff after a reviewer-supplied link enters the ingestion workflow.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "file_name": {
            "title": "File Name",
            "type": "string"
          },
          "host": {
            "title": "Host",
            "type": "string"
          },
          "ingestion_job_id": {
            "title": "Ingestion Job Id",
            "type": "string"
          },
          "ingestion_status": {
            "title": "Ingestion Status",
            "type": "string"
          },
          "source_format": {
            "$ref": "#/components/schemas/FileFormat"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "host",
          "file_name",
          "source_format",
          "ingestion_job_id",
          "ingestion_status",
          "created_at"
        ],
        "title": "LinkAcquisitionStart",
        "type": "object"
      },
...
      "SourceAcquisitionRequest": {
        "description": "Reviewer identity selecting one candidate previously returned by the agent.",
        "properties": {
          "candidateId": {
            "pattern": "^[a-z][a-z0-9_-]*$",
            "title": "Candidateid",
            "type": "string"
          },
          "submittedBy": {
            "maxLength": 320,
            "minLength": 1,
            "title": "Submittedby",
            "type": "string"
          }
        },
        "required": [
          "candidateId",
          "submittedBy"
        ],
        "title": "SourceAcquisitionRequest",
        "type": "object"
      },
      "SourceCandidate": {
        "description": "Allowlisted source metadata safe to expose to a reviewer.",
        "properties": {
          "candidate_id": {
            "pattern": "^[a-z][a-z0-9_-]*$",
            "title": "Candidate Id",
            "type": "string"
          },
          "connector_id": {
            "pattern": "^[a-z][a-z0-9_-]*$",
            "title": "Connector Id",
            "type": "string"
          },
          "download_url": {
            "pattern": "^https://",
            "title": "Download Url",
            "type": "string"
          },
          "entity_ids": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Entity Ids",
            "type": "array"
          },
          "file_name": {
            "minLength": 1,
            "title": "File Name",
            "type": "string"
          },
          "license": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License"
          },
          "metric_codes": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Metric Codes",
            "type": "array"
          },
          "period_end": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period End"
          },
          "period_start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Start"
          },
          "publisher": {
            "minLength": 1,
            "title": "Publisher",
            "type": "string"
          },
          "source_format": {
            "$ref": "#/components/schemas/FileFormat"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "topic_terms": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Topic Terms",
            "type": "array"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "candidate_id",
          "connector_id",
          "title",
          "publisher",
          "download_url",
          "file_name",
          "source_format"
        ],
        "title": "SourceCandidate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "New Taipei Youth Policy API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/copilot/acquisitions": {
      "post": {
        "description": "Snapshot an allowlisted candidate into the approval-gated ingestion workflow.",
        "operationId": "acquire_source_api_v1_copilot_acquisitions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceAcquisitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcquisitionStart"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Acquire Source",
        "tags": [
          "copilot"
        ]
      }
    },
    "/api/v1/copilot/acquisitions/link": {
      "post": {
        "description": "Snapshot a reviewer link from an approved host into the approval-gated workflow.",
        "operationId": "acquire_link_api_v1_copilot_acquisitions_link_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkAcquisitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkAcquisitionStart"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Acquire Link",
        "tags": [
          "copilot"
        ]
      }
    },
...
    "/api/v1/copilot/query": {
      "post": {
        "description": "Execute an allowlisted decision plan and return grounded evidence.",
        "operationId": "query_copilot_api_v1_copilot_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CopilotQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CopilotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Query Copilot",
        "tags": [
          "copilot"
        ]
      }
    },
    "/api/v1/copilot/query/stream": {
      "post": {
        "description": "Stream workflow stages and Bedrock deltas, then the validated response.",
        "operationId": "stream_copilot_api_v1_copilot_query_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CopilotQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Copilot",
        "tags": [
          "copilot"
        ]
      }
    },
...
  }
}